requirements.txt
gradio==4.37.2
matplotlib==3.9.0
mmcls==0.25.0
mmcv==2.2.0
mmdeploy==1.3.1
mmdet==3.3.0
mmengine==0.10.4
mmpose==1.3.1
mmrazor==1.0.0
mmrotate==0.3.4
numpy==1.23.5
onnx==1.14.0
onnxruntime==1.15.1
onnxruntime_gpu==1.18.0
onnxsim==0.4.36
opencv_python==4.10.0.82
pandas==2.2.2
parameterized==0.9.0
Pillow==10.4.0
prettytable==3.10.0
pycocotools==2.0.7
pytest==8.2.2
pytorch_sphinx_theme==0.0.19
PyYAML==6.0.1
PyYAML==6.0.1
rich==13.7.1
sahi==0.11.16
scipy==1.14.0
seaborn==0.13.2
setuptools==70.0.0
Shapely==2.0.4
supervision==0.21.0
tensorrt==10.1.0
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tqdm==4.66.4
